+++ CHANGELOG.md
@@ -64,6 +64,14 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
   but please report any regressions that you may find.
 - The dot access ellipsis now matches field accesses in addition to method
   calls.
+- pattern-regex, pattern-not-regex, metavariable-regex: `^` and `$`
+  now match at the beginning and end of each line, respectively,
+  rather than previously just at the beginning and end of the input
+  file. This corresponds to PCRE's multiline mode. To get the old
+  behavior back, use `\A` instead of '^' and `\Z` instead of `$`. See
+  the [PCRE
+  manual](https://www.pcre.org/original/doc/html/pcrepattern.html#smallassertions)
+  for details.
 - Made error message for resource exhausion (exit code -11/-9) more actionable
 - Made error message for rules with patterns missing positive terms
   more actionable (#5234)